+++ entity/src/epilot/models/shared/consentattribute.py
@@ -3,8 +3,9 @@
 from __future__ import annotations
 import dataclasses
 from dataclasses_json import Undefined, dataclass_json
+from enum import Enum
 from epilot import utils
-from typing import Any, Final, Optional
+from typing import Any, Optional
 
 
 
@@ -42,6 +43,9 @@ class ConsentAttributeInfoHelpers:
     
 
 
+class ConsentAttributeType(str, Enum):
+    CONSENT = 'consent'
+
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
@@ -51,7 +55,7 @@ class ConsentAttribute:
     label: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('label') }})
     name: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('name') }})
     topic: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('topic') }})
-    TYPE: Final[str] = dataclasses.field(default='consent', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: ConsentAttributeType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     purpose: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('_purpose'), 'exclude': lambda f: f is None }})
     constraints: Optional[ConsentAttributeConstraints] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('constraints'), 'exclude': lambda f: f is None }})
     r"""A set of constraints applicable to the attribute.