+++ entity/src/epilot/models/shared/currencyattribute.py
@@ -3,8 +3,9 @@
 from __future__ import annotations
 import dataclasses
 from dataclasses_json import Undefined, dataclass_json
+from enum import Enum
 from epilot import utils
-from typing import Any, Final, Optional, Union
+from typing import Any, Optional, Union
 
 
 
@@ -61,6 +62,9 @@ class CurrencyAttributeInfoHelpers:
     
 
 
+class CurrencyAttributeType(str, Enum):
+    CURRENCY = 'currency'
+
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
@@ -71,7 +75,7 @@ class CurrencyAttribute:
     r"""An array of currency configurations with a country code (ISO-4217)"""
     label: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('label') }})
     name: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('name') }})
-    TYPE: Final[str] = dataclasses.field(default='currency', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: CurrencyAttributeType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     purpose: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('_purpose'), 'exclude': lambda f: f is None }})
     constraints: Optional[CurrencyAttributeConstraints] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('constraints'), 'exclude': lambda f: f is None }})
     r"""A set of constraints applicable to the attribute.