+++ entity/src/epilot/models/shared/statusattribute.py
@@ -3,8 +3,9 @@
 from __future__ import annotations
 import dataclasses
 from dataclasses_json import Undefined, dataclass_json
+from enum import Enum
 from epilot import utils
-from typing import Any, Final, Optional, Union
+from typing import Any, Optional, Union
 
 
 
@@ -60,6 +61,9 @@ class StatusAttributeOptions2:
 class StatusAttributeOptions:
     pass
 
+class StatusAttributeType(str, Enum):
+    STATUS = 'status'
+
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
@@ -112,7 +116,7 @@ class StatusAttribute:
     r"""Render as a column in table views. When defined, overrides `hidden`"""
     sortable: Optional[bool] = dataclasses.field(default=True, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('sortable'), 'exclude': lambda f: f is None }})
     r"""Allow sorting by this attribute in table views if `show_in_table` is true"""
-    TYPE: Final[Optional[str]] = dataclasses.field(default='status', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})
+    type: Optional[StatusAttributeType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})
     value_formatter: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('value_formatter'), 'exclude': lambda f: f is None }})